added useJournal, migrated relevant functions from JournalsCtx & JournalsPage to useJournal
